pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audiobook-converter"
version = "0.1.0"
description = "Intelligent novel to audiobook converter with character voice matching"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "pydub>=0.25.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
audiobook = "audiobook.cli:main"

[tool.setuptools.packages.find]
where = ["src"]